feat: ajout de la vue Semaine et composants associés pour le chargement et l'absence de cours
