+++ src/consts/DefaultTabs.ts
@@ -63,5 +63,12 @@ export const defaultTabs = [
     description: "Tes compétences et évaluations",
     icon: require("@/../assets/lottie/tab_evaluations.json"),
     enabled: true,
-  }
+  },
+  {
+    tab: "Week",
+    label: "Semaine",
+    description: "Vue avancée de la semaine",
+    icon: require("@/../assets/lottie/tab_calendar.json"),
+    enabled: true,
+  },
 ] as const;
