+++ src/router/screens/views/index.ts
@@ -11,8 +11,8 @@ import RestaurantHistory from "@/views/account/Restaurant/Modals/History";
 import ChatCreate from "@/views/account/Chat/Modals/ChatCreate";
 import Chat from "@/views/account/Chat/Modals/Chat";
 import HomeworksDocument from "@/views/account/Homeworks/Document";
-import LessonsImportIcal from "@/views/account/Lessons.old/Options/LessonsImportIcal";
-import LessonDocument from "@/views/account/Lessons.old/Document";
+import LessonsImportIcal from "@/views/account/Lessons/Options/LessonsImportIcal";
+import LessonDocument from "@/views/account/Lessons/Document";
 import BackgroundIUTLannion from "@/views/login/IdentityProvider/actions/BackgroundIUTLannion";
 import { Platform } from "react-native";
 import GradeReaction from "@/views/account/Grades/Modals/GradeReaction";
